backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.3
scipy==1.12.0
uuid
gunicorn==21.2.0
werkzeug==3.0.1